+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0 - 2021-02-19
+
+### Added
+
+- Added a `config :harness, :version` key which checks the version of harness
+  and ensures that a manifest is only rendered if the harness version matches
+
 ## 0.3.0 - 2021-02-17
 
 ### Added